Refactor the details DID table views
