--- src/component-library/pages/DID/details/DetailsDIDComponent.ts
@@ -1,6 +1,0 @@
-export type DetailsDIDProps = {
-    scope: string;
-    name: string;
-};
-
-export type DetailsDIDComponent = (props: DetailsDIDProps) => JSX.Element;
