+++ src/component-library/pages/DID/details/tables/DetailsDIDSimpleTable.tsx
