+++ src/component-library/pages/DID/details/views/DetailsDIDContents.tsx
@@ -1,15 +1,15 @@
 import { DIDViewModel } from '@/lib/infrastructure/data/view-model/did';
 import React, { useEffect } from 'react';
-import { DetailsDIDComponent, DetailsDIDProps } from '@/component-library/pages/DID/details/DetailsDIDComponent';
+import { DetailsDIDView, DetailsDIDProps } from '@/component-library/pages/DID/details/views/DetailsDIDView';
 import useTableStreaming from '@/lib/infrastructure/hooks/useTableStreaming';
-import { DetailsDIDSimpleTable } from '@/component-library/pages/DID/details/DetailsDIDSimpleTable';
+import { DetailsDIDSimpleTable } from '@/component-library/pages/DID/details/tables/DetailsDIDSimpleTable';
 
-export const DetailsDIDParents: DetailsDIDComponent = ({ scope, name }: DetailsDIDProps) => {
+export const DetailsDIDContents: DetailsDIDView = ({ scope, name }: DetailsDIDProps) => {
     const { gridApi, onGridReady, streamingHook, startStreaming, stopStreaming } = useTableStreaming<DIDViewModel>();
 
     useEffect(() => {
         if (gridApi) {
-            const url = '/api/feature/list-did-parents?' + new URLSearchParams({ scope, name });
+            const url = '/api/feature/list-did-contents?' + new URLSearchParams({ scope, name });
             startStreaming(url);
         }
     }, [gridApi]);
