+++ src/component-library/pages/DID/details/views/DetailsDIDView.ts
@@ -1,0 +1,6 @@
+export type DetailsDIDProps = {
+    scope: string;
+    name: string;
+};
+
+export type DetailsDIDView = (props: DetailsDIDProps) => JSX.Element;
